examples/socketio-webcam-videostream/pycli/requirements.txt
aiortc
aiohttp
av
python-socketio[asyncio-client]
asyncio
opencv-python
